2.2.0.2 Emergency Update
An attacker discovered a vulnerability that allowed them to build off blocks deeper in the chain than besthight and have them be accepted by the network.

This update adds the following consensus rules-

A new block can only orphan our current best chain blocks if:

The blocks that it is attempting to displace are less than 2 minutes old.

OR

ALL of the blocks that it is attempting to displace have less than ideal spacing, relative to each other.

AND

The time spacing between our best block and the block that they are attempting to build on is less than ideal limits.

This means that if there's a flood of blocks then they're up for contention until someone builds a suitibly spaced block on top of them that is accepted by the chain for at least 2 minutes without contention.

Note that this introduces a level of chain trust that we did not previously have - we can now analyze our best chain and determine with certainty at what block transactions should be considered absolutely and inalterably confirmed by the network.

Note: As the network grows we may need to increase the 2 minute threshold in order to maintain stability, but at present it is well within tolerable limits for a well connected wallet or node.