Allow Checkpoint Override on 2.2.0.2
Non-critical/Non-mandatory update.

Update to this if you get stuck and unable to sync after the 2.2.0.2 update goes into effect.

There is an unlikely but possible scenario where you could end up with a valid block that meets all of the 2.2.0.2 criteria but the network has elected to follow an alternative chain to the one that you have. This update will allow you to sync to the most recent network checkpoint.

This should not be necessary under most circumstances, but exists as a precaution in case forks do occur.